# frontend/src/api/openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/studios":{"get":{"operationId":"StudiosController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Studios"]},"post":{"operationId":"StudiosController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateStudioDto"}}}},"responses":{"201":{"description":""}},"tags":["Studios"]}},"/api/studios/{id}":{"patch":{"operationId":"StudiosController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStudioDto"}}}},"responses":{"200":{"description":""}},"tags":["Studios"]}},"/api/studios/{studioId}/packages":{"get":{"operationId":"PackagesController_findAll","parameters":[{"name":"studioId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Packages"]},"post":{"operationId":"PackagesController_create","parameters":[{"name":"studioId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePackageDto"}}}},"responses":{"201":{"description":""}},"tags":["Packages"]}},"/api/studios/{studioId}/packages/{id}":{"patch":{"operationId":"PackagesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePackageDto"}}}},"responses":{"200":{"description":""}},"tags":["Packages"]},"delete":{"operationId":"PackagesController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Packages"]}},"/api/studios/{studioId}/sessions":{"post":{"operationId":"SessionsController_create","parameters":[{"name":"studioId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSessionDto"}}}},"responses":{"201":{"description":""}},"tags":["Sessions"]},"get":{"operationId":"SessionsController_findByStudio","parameters":[{"name":"studioId","required":true,"in":"path","schema":{"type":"string"}},{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/api/sessions/{id}":{"get":{"operationId":"SessionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/api/sessions/validate":{"get":{"operationId":"SessionsController_validateToken","parameters":[{"name":"token","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/api/sessions/{sessionId}/confirm":{"post":{"operationId":"SessionsController_confirm","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-access-token","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Sessions"]}},"/api/studios/{studioId}/sessions/{sessionId}/status":{"patch":{"operationId":"SessionsController_updateStatus","parameters":[{"name":"studioId","required":true,"in":"path","schema":{"type":"string"}},{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSessionStatusDto"}}}},"responses":{"200":{"description":""}},"tags":["Sessions"]}},"/api/sessions/{sessionId}/photos":{"post":{"operationId":"PhotosController_uploadPhotos","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Photos"]},"get":{"operationId":"PhotosController_findAll","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Photos"]}},"/api/sessions/{sessionId}/photos/sign":{"get":{"operationId":"PhotosController_getUploadSignature","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Photos"]}},"/api/sessions/{sessionId}/photos/confirm":{"post":{"operationId":"PhotosController_confirmUpload","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmUploadDto"}}}},"responses":{"201":{"description":""}},"tags":["Photos"]}},"/api/sessions/{sessionId}/photos/{photoId}/select":{"post":{"operationId":"PhotosController_select","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}},{"name":"photoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-access-token","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Photos"]}},"/api/sessions/{sessionId}/photos/{photoId}/deselect":{"post":{"operationId":"PhotosController_deselect","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}},{"name":"photoId","required":true,"in":"path","schema":{"type":"string"}},{"name":"x-access-token","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Photos"]}},"/api/studios/{studioId}/sessions/{sessionId}/selected":{"get":{"operationId":"StudioSessionPhotosController_getSelected","parameters":[{"name":"sessionId","required":true,"in":"path","schema":{"type":"string"}},{"name":"export","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["StudioSessionPhotos"]}}},"info":{"title":"MiSesionPro API","description":"Documentación de la API de MiSesionPro","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"RegisterDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}},"CreateStudioDto":{"type":"object","properties":{}},"UpdateStudioDto":{"type":"object","properties":{}},"CreatePackageDto":{"type":"object","properties":{}},"UpdatePackageDto":{"type":"object","properties":{}},"CreateSessionDto":{"type":"object","properties":{}},"UpdateSessionStatusDto":{"type":"object","properties":{}},"ConfirmUploadDto":{"type":"object","properties":{}}}}}